--- a/test/support/fixture_methods.rb
+++ b/test/support/fixture_methods.rb
@@ -7,22 +7,24 @@ module FixtureMethods
     File.read(fixture_file(filename))
   end
 
+  def fixture_flows_path
+    Rails.root.join("test/fixtures/smart_answer_flows")
+  end
+
+  def require_fixture_flows
+    Dir[fixture_flows_path.join("*.rb")].map { |path| require path }
+  end
+
   def setup_fixture_flows
     stub_request(:get, %r{#{Plek.new.find("content-store")}/content/(.*)})
       .to_return(status: 404, body: {}.to_json)
 
-    fixture_flows_path = Rails.root.join("test/fixtures/smart_answer_flows")
-    # require each of the flows since they're outside the autload path
-    Dir[fixture_flows_path.join("*.rb")].map { |path| require path }
+    require_fixture_flows
 
-    @preload_flows = SmartAnswer::FlowRegistry.instance.preloaded?
-    SmartAnswer::FlowRegistry.reset_instance(
-      preload_flows: false,
-      smart_answer_load_path: fixture_flows_path,
-    )
+    SmartAnswer::FlowRegistry.reset_instance(smart_answer_load_path: fixture_flows_path)
   end
 
   def teardown_fixture_flows
-    SmartAnswer::FlowRegistry.reset_instance(preload_flows: @preload_flows)
+    SmartAnswer::FlowRegistry.reset_instance
   end
 end