--- a/test/unit/flow_registry_test.rb
+++ b/test/unit/flow_registry_test.rb
@@ -2,6 +2,8 @@ require_relative "../test_helper"
 
 module SmartAnswer
   class FlowRegistryTest < ActiveSupport::TestCase
+    setup { require_fixture_flows }
+
     def registry(options = {})
       FlowRegistry.new(options.merge(smart_answer_load_path: File.expand_path("../fixtures/smart_answer_flows", __dir__)))
     end
@@ -28,42 +30,6 @@ module SmartAnswer
       flows = registry.flows
       assert_kind_of Enumerable, flows
       assert_kind_of Flow, flows.first
-    end
-
-    context "without preloaded flows" do
-      setup do
-        @r = registry
-      end
-
-      should "build a new flow instance each time" do
-        first_call = registry.find("flow-sample")
-        second_call = registry.find("flow-sample")
-
-        assert_equal first_call.name, second_call.name
-        assert_not_equal first_call.object_id, second_call.object_id
-      end
-    end
-
-    context "with preloaded flows" do
-      setup do
-        @r = registry(preload_flows: true)
-      end
-
-      should "not hit the filesystem when finding a flow" do
-        Dir.expects(:[]).never
-        File.expects(:read).never
-
-        assert @r.find("flow-sample").is_a?(Flow)
-      end
-
-      should "not hit the filesystem when finding a non-existent flow" do
-        Dir.expects(:[]).never
-        File.expects(:read).never
-
-        assert_raises FlowRegistry::NotFound do
-          @r.find("non_existent")
-        end
-      end
     end
   end
 end