Trello: https://trello.com/c/fU4RQYvF/494-foundational-work-to-implement-smart-answer-flow-test-adr

Avoiding preloading of flows in development has a near negligible impact
on performance compared to more significant things (notably view
rendering). However it's presence means that we need a quirky system
where Flow classes have an unconventional file name pattern where that
needs special autoload code and the extra logic in this flow registry
class.

This has been removed as a precursor to a step where the class names of
Flows are changed from their unusual kebab-case style (e.g.
benefits-cap-calculator.rb) to the conventional Ruby snake_case style
(e.g. benefits_cap_calculator_flow.rb).

Some benchmarks from rendering /benefits-cap-calculator

main branch, initial request: Completed 200 OK in 2040ms (Views: 1934.1ms | Allocations: 1066068)
main branch, subsequent request: Completed 200 OK in 725ms (Views: 675.3ms | Allocations: 169443)

this branch, initial request: Completed 200 OK in 2100ms (Views: 1889.1ms | Allocations: 1136213)
this branch, subsequent request: Completed 200 OK in 466ms (Views: 410.7ms | Allocations: 168485)

The impact of this change is that there are more object allocations on
an initial request, however given the view rendering takes ~2s and
the time to load flows takes ~0.1s it's unlikely to be noticeable
for any developers.

⚠️ This repo is Continuously Deployed: make sure you [follow the guidance](https://docs.publishing.service.gov.uk/manual/development-pipeline.html#merge-your-own-pull-request) ⚠️